requirements.txt
tk
Pillow
pywin32
